INCOMPLETE -- enable position independent code
This is kinda crippled, but it at least resembles PIC...

Globals are accessed using "direct" addressing.  This limits us to 256
bytes of global storage, but at least it should be usable to generate
OS-9 executables (with the proper startup code and image generation).

Still need to resolve the options handling in the "bcc" driver binary.