bcc: remove still more unused (mostly x86) stuff
This also removed opt_arch completely, although it may need to come back
in some form for supporting multiple environments (e.g. DECB, OS-9,
FLEX, stand-alone, etc) and/or if HD6309 support is added in the future.

Also, there remains a bit of stuff that seems to be intended either for
running bcc09 as a native compiler (i.e. hosted on the target) and/or
for compiling bcc09 with itself...

Signed-off-by: John W. Linville <linville@tuxdriver.com>