objcopy: add '-p' option for use with binary output type
This option makes the binary output from objcopy match what is produced
by the '-d' option to ld when a '-D' option is used to offset the
data section from the end of the text section.

Also, add a note to the TODO file about making ld and objcopy more
sanely handle a -D argument that is less than the -T argument.

Signed-off-by: John W. Linville <linville@tuxdriver.com>